Fixes #78 and reorganizes sandbox in general.
Moved code around, all sandbox APIs are now in the configuration
project, the implementations are still in js/sandbox.

Zed-specific files now start with `zed::` instead of `zed:` and appear
in the goto file list so you can always go to `zed::start` to find the
cheat sheet.

The sandbox now runs in a Chrome webview, which runs in a separate
process therefore no longer blocking the Zed UI thread.