Add newline when saving user prefs.
Editing `user.json` by hand results in the whitespace stripper ensuring a newline at the end of the file. But if you use a command to change a setting and zed writes out a JSON dump of `user.json` for you, it lacks the trailing newline. This annoying inconsistency makes it difficult to keep `user.json` in version control, because sometimes it has the trailing newline and sometimes it doesn't.

This change ensures the `user.json` will always have the trailing newline.